removed unwated interfaces; renamed publisher to producer and reimplement send method on producer;
