+++ packages/infratructure/src/index.ts
@@ -42,5 +42,5 @@
 
 export * from "./oracles/oracleprovider";
 export * from "./oracles/oraclefinder";
-export * from "./messagepublisher";
+export * from "./messageproducer";
 export * from "./localcache";
